pyproject.toml
[project]
name = "predictive-poultry-systems"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "Pillow>=10.0.0",
    "faker>=40.12.0",
    "greenlet>=3.3.2",
    "prek>=0.3.8",
    "pydantic>=2.12.5",
    "ruff>=0.15.9",
    "salabim>=24.0.12",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.1.0",
]
